Move Node-/RelationshipTable to okapi-relational
