+++ okapi-relational/src/main/scala/org/opencypher/okapi/relational/api/table/RelationalCypherRecords.scala
@@ -27,6 +27,8 @@
 package org.opencypher.okapi.relational.api.table
 
 import org.opencypher.okapi.api.table.CypherRecords
+import org.opencypher.okapi.impl.table.RecordsPrinter
+import org.opencypher.okapi.impl.util.PrintOptions
 import org.opencypher.okapi.relational.api.io.EntityTable
 import org.opencypher.okapi.relational.impl.table.RecordHeader
 
@@ -53,5 +55,10 @@ trait RelationalCypherRecords[T <: Table[T]] extends CypherRecords {
 
   def cache(): Records
 
+  override def size: Long = table.size
+
   override def physicalColumns: Seq[String] = table.physicalColumns
+
+  override def show(implicit options: PrintOptions): Unit =
+    RecordsPrinter.print(this)
 }
